Split out client/server parts from bogo::main
